WiFiDisplayController: Defer the P2P Initialization from its constructor.
P2PManager initialization from WiFiDisplayController’s constructor
would start a p2p interface on the boot up when persist.debug.wfd.enable is set.
This would block the NAN operations if NAN and P2P concurrency is not
supported.
Moreover, this initialization of P2P in WiFiDisplayController’s constructor is
needless as this would be required for any of the P2P operations that would
happen at later point of time, based on the user’s trigger.
Thus, move this P2P initialization to the first action point of P2P operation
in WiFiDisplayController.

NOTE: Once initialization of P2P is done in WiFiDisplayController, it is not
de-initialized again. Thus P2P interface would exist and further NAN start would
fail until device is rebooted. This issue need to be fixed.

Change-Id: Ia26aee2f6496cf11aba3c99dcb55c3264ea0a5c0
CRs-Fixed: 2258826
Signed-off-by: drpradsms <dr.pradsms@gmail.com>